Fix model export. Fix leaks. Fix replace marksurf
Fix model export. Fix leaks. Fix replace zero marksurf index. (missing and crash map sometime). Bug fixes.
Added new crash after remove internal textures.

+++ src/editor/LightmapNode.h
@@ -5,7 +5,7 @@ class LightmapNode
 {
 public:
 	LightmapNode* child[2];
-	short x, y, w, h;
+	int x, y, w, h;
 	bool filled;
 
 	LightmapNode(int offX, int offY, int mapW, int mapH);
